Extract retry logic to RetryManager
